% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchRateMatrix.R
\name{branchRateMatrix}
\alias{branchRateMatrix}
\title{Compute exponentiated rate matrix for all branches}
\usage{
branchRateMatrix(rate, branch.length, pi, log = TRUE)
}
\arguments{
\item{rate}{vector of mutation rates, one per branch}

\item{branch.length}{vector of branch lengths}

\item{pi}{stationary distribution of allele frequencies}

\item{log}{if TRUE returns log transition matricies}
}
\value{
a list of transition matricies, one entry for each specified branch length. The rows are the parental allele,
the columns are the child allele.
}
\description{
Compute exponentiated rate matrix for all branches
}
