% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchTransitions.R
\name{branchTransitions}
\alias{branchTransitions}
\title{Compute transitions per branch}
\arguments{
\item{obj}{rateModel object}

\item{subset}{The site indices for which should be summed over to compute the total number of expected transitions (NOT IMPLEMENTED)}

\item{ncores}{Number of cores to use}
}
\value{
A 3-dimensional vectors where the first index is the branch, the second is the state of the parent, and the third 
is the state of the child. The entries are the exprected number of transitions on that branch for the corresponding parent
and child states.
}
\description{
Compute the number of expected transitions per branch between states based on the marginal joint probabilities
}
\examples{

}
