% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fits rate model}
\arguments{
\item{obj}{rateModel}

\item{scale}{a scale factor to apply to log-likelihood, defaults to -1/nsites}
}
\value{
rateModel object
}
\description{
Fits rate model object and returns fitted model
}
\examples{

}
