% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdgeGroup.R
\name{getEdgeGroup}
\alias{getEdgeGroup}
\title{Get edgeGroup}
\arguments{
\item{obj}{rateModel object}

\item{parent}{node number(s) for parents}

\item{child}{node number(s) for children}
}
\description{
Returns edgeGroups table as data.table
}
\examples{

}
