% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdgeTable.R
\name{getEdgeTable}
\alias{getEdgeTable}
\title{Tree visualization}
\arguments{
\item{x}{alleleData}
}
\value{
data.table containing edge ids
}
\description{
Plots tree with various helpful highlightings
}
\examples{

}
