% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndexLabels.R
\name{getIndexLabels}
\alias{getIndexLabels}
\title{Get all site labels for an index}
\arguments{
\item{obj}{a rateModel object}

\item{i}{a vector of site indices}
}
\value{
a factor vector of labels
}
\description{
Gets label of all sites with a given index
}
\examples{

}
