% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabelIndices.R
\name{getLabelIndices}
\alias{getLabelIndices}
\title{Get all site indices for a label}
\arguments{
\item{obj}{a rateModel object}

\item{label}{a site label}
}
\value{
a numeric vector of indices
}
\description{
Gets indicies of all sites with a given label
}
\examples{

}
