% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamValue.R
\name{getParamValue}
\alias{getParamValue}
\title{Get parameter values}
\arguments{
\item{obj}{rateModel object}

\item{i}{a numeric vector of indices}
}
\description{
Get parameter values given a set of indices
}
\examples{

}
