% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamIndex.R
\name{getParamIndex}
\alias{getParamIndex}
\title{Get parameter index}
\arguments{
\item{obj}{rateModel object}
}
\value{
A data.table of parameter indices, keyed by edgeGroups and siteLabels
}
\description{
Returns parameter index as data.table
}
\examples{

}
