% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRateIndex.R
\name{getRateIndex}
\alias{getRateIndex}
\title{Get rate parameter index}
\arguments{
\item{obj}{rateModel object}

\item{edges}{Can either be a data.frame/data.table with two columns, \'parent\' and \'child\' or a character vector of edgeIDs}

\item{siteLabel}{A character vector of length one for one of the site labels}

\item{edgeGroup}{Instead of the edges argument, a vector of edge groups can be supplied. Only one can be specified.}
}
\value{
A numeric vector of parameter indices
}
\description{
Returns index of a rate parameter associated with a given edge/edgeGroup and siteLabel
}
\examples{

}
