% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSibbings.R
\name{getSibblings}
\alias{getSibblings}
\title{Compute transitions per branch}
\usage{
getSibblings(tr, node)
}
\arguments{
\item{tr}{phylo object}
}
\value{
The node numbers of sibbling nodes
}
\description{
Compute the number of expected transitions per branch between states based on the marginal joint probabilities
}
\examples{

}
