% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteInfo.R
\name{getSiteInfo}
\alias{getSiteInfo}
\title{Extract site information table}
\arguments{
\item{obj}{alleleData or rateModel object}
}
\value{
data.table containing site info
}
\description{
Extracts site information table
}
\examples{

}
