% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikelihood.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Compute log-likelihood}
\arguments{
\item{obj}{rateModel object}
}
\value{
A rateModel object with updated parameter values
}
\description{
Compute log-likelihood for given rate model object
}
\examples{

}
