% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiProbToStick}
\alias{multiProbToStick}
\title{multiProbToStick}
\usage{
multiProbToStick(x, width)
}
\arguments{
\item{x}{the parameter of stick breaking parameters}

\item{width}{the width of the stick breakink process}
}
\value{
a vector of probabilities that sum to one
}
\description{
Convert a multiple sets of stick breaking parameters to a set of probabilities that sum to one
}
