% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\title{Tree visualization}
\arguments{
\item{obj}{alleleData}

\item{colorByRate}{Can be set to either color by the grouping of rates or their actual values. Only usable with rateModel object.}
}
\description{
Plots tree with various helpful highlightings
}
\examples{

}
