% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probToStick}
\alias{probToStick}
\title{probToStick}
\usage{
probToStick(x)
}
\arguments{
\item{x}{parameter vector of probabilities}
}
\value{
a vector of parameters for a stick breaking parameters
}
\description{
Convert a set of probabilities that sum to one to a set of stick breaking parameters
}
