% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateSE.R
\name{rateSE}
\alias{rateSE}
\title{Estimates the standard error of the rate}
\arguments{
\item{obj}{rateModel}
}
\value{
a data.table with the parameter values and standard errors
}
\description{
Estimates the standard error for all rates in the model using the hessian
}
\examples{

}
