% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledLL.R
\name{scaledLL}
\alias{scaledLL}
\title{Compute scaled negative log-likelihood}
\arguments{
\item{x}{a set of parameters to compute the log-likelihood for}

\item{obj}{rateModel object}
}
\value{
A rateModel object with updated parameter values
}
\description{
Compute scaled negative log-likelihood for given rate model object
}
