% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParamValue.R
\name{setParamValue}
\alias{setParamValue}
\title{Set parameter values}
\arguments{
\item{obj}{rateModel object}

\item{i}{a numeric vector of indices to set to \'value\' argument}

\item{value}{a numeric vector of values to set}
}
\value{
A rateModel object with updated parameter values
}
\description{
Get parameter values given a set of indices
}
