% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulate data}
\usage{
simData(nSites, tr, rate, pi)
}
\arguments{
\item{nSites}{number of sites to simulate}

\item{tr}{tree}

\item{rate}{a vector, either of length one, or the same as the number of branches}

\item{pi}{stationary distribution of allele frequencies}
}
\value{
a list
}
\description{
Simulates data without an error model
}
