% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteTransitions.R
\name{siteTransitions}
\alias{siteTransitions}
\title{Compute transitions per branch}
\arguments{
\item{obj}{rateModel object}

\item{subset}{The site indices for which should be summed over to compute the total number of expected transitions (NOT IMPLEMENTED)}

\item{ncores}{Number of cores to use}
}
\value{
A data.table where each is labeled by site index, parent allele, and child allele, and the number transitions over the whole tree
}
\description{
Compute the number of expected transitions per site between states based on the marginal joint probabilities
}
\examples{

}
