% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stickToProb}
\alias{stickToProb}
\title{stickToProb}
\usage{
stickToProb(x)
}
\arguments{
\item{x}{parameter vector of stick breaking process parameters}
}
\value{
a vector of probabilities that sum to one
}
\description{
Convert a set of stick breaking parameters to a set of probabilities that sum to one
}
