% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tieRates.R
\name{tieRates}
\alias{tieRates}
\title{Tie rates across siteLabels or edgeGroups}
\arguments{
\item{obj}{a rateModel object}

\item{siteLabelA}{}

\item{edgeGroupA}{foo}

\item{siteLabelB}{a label}

\item{edgeGroupB}{foo}
}
\description{
Sets the rate of the set of branches specified by edgeGroupB and siteLabelB equal to that of A
}
\examples{

}
