% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fits rate model}
\arguments{
\item{obj}{rateModel}

\item{scale}{a scale factor to apply to log-likelihood, defaults to -1/nsites}

\item{method}{Optimization method to use ("l-bfgs-b","mlsl")}
}
\value{
a list incuding a rateModel object and information about the optimization
}
\description{
Fits rate model object and returns fitted model
}
\examples{

}
