% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdgeTable.R
\name{getEdgeTable}
\alias{getEdgeTable}
\title{Get edge table}
\arguments{
\item{x}{alleleData}
}
\value{
data.table containing edge ids
}
\description{
Returns edge table. Useful for construction of rateModel class.
}
\examples{

}
