% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdgeTable.R, R/getSiteInfo.R
\name{getEdgeTable}
\alias{getEdgeTable}
\title{Tree visualization}
\arguments{
\item{x}{alleleData}

\item{x}{alleleData or rateModel object}
}
\value{
data.table containing edge ids

data.table containing edge ids
}
\description{
Plots tree with various helpful highlightings

Extracts site information table
}
\examples{


}
