% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPiIndex.R
\name{getPiIndex}
\alias{getPiIndex}
\title{Get pi parameter index}
\arguments{
\item{obj}{rateModel object}

\item{siteLabel}{A character vector of length one for one of the site labels}
}
\value{
A numeric vector of parameter indices
}
\description{
Returns indecies of the stationary distribution parameters associated with a given siteLabel
}
\examples{

}
