% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRateIndex.R
\name{getRateIndex}
\alias{getRateIndex}
\title{Get parameter index}
\arguments{
\item{obj}{rateModel object}

\item{edges}{Can either be a data.frame/data.table with two columns, \'parent\' and \'child\' or a character vector of edgeIDs}

\item{siteLabel}{A character vector of length one for one of the site labels}
}
\value{
A numeric vector of parameter indices
}
\description{
Returns parameter index as data.table
}
\examples{

}
