% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateModel-class.R
\name{rateModel}
\alias{rateModel}
\title{rateModel}
\usage{
rateModel(data, siteLabelCriteria = NULL, lineageTable = NULL)
}
\arguments{
\item{data}{An alleleData object}

\item{siteLabelCriteria}{A charecter vector of the columns in siteData contained in alleleData label genomic regions}

\item{lineageTable}{A table with columns parent,child,edgeID, and rateGroup, where rateGroup is used to specify how the rates are tied between the branches}
}
\value{
an rateModel object
}
\description{
Contructs an object that holds the allele data and the phylogenetic tree
}
\examples{

}
