% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tieRates.R
\name{tieRates}
\alias{tieRates}
\title{Tie rates across siteLabels or edgeGroups}
\arguments{
\item{obj}{a rateModel object}

\item{siteLabelA}{site label for rate A}

\item{edgeGroupA}{edge group for rate A}

\item{siteLabelB}{site label for rate B}

\item{edgeGroupB}{edge group for rate B}
}
\description{
Sets the rate of the set of branches specified by edgeGroupB and siteLabelB equal to that of A
}
\examples{

}
