% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tieRates.R
\name{tieRates}
\alias{tieRates}
\title{Get all site indices for a label}
\arguments{
\item{obj}{a rateModel object}

\item{siteLabelA}{a label}

\item{edgeGroupA}{foo}

\item{siteLabelB}{a label}

\item{edgeGroupB}{foo}
}
\description{
Gets indicies of all sites with a given label
}
\examples{

}
